Make flaky tracking test a ui test
